Bulk Workflow Manager v2 - Code review alignment
