+++ bundle/src/main/java/com/adobe/acs/commons/workflow/bulk/execution/BulkWorkflowRunner.java
@@ -22,7 +22,7 @@ package com.adobe.acs.commons.workflow.bulk.execution;
 
 import aQute.bnd.annotation.ProviderType;
 import com.adobe.acs.commons.util.QueryHelper;
-import com.adobe.acs.commons.workflow.bulk.execution.impl.SubStatus;
+import com.adobe.acs.commons.workflow.bulk.execution.model.SubStatus;
 import com.adobe.acs.commons.workflow.bulk.execution.model.Config;
 import com.adobe.acs.commons.workflow.bulk.execution.model.Payload;
 import com.adobe.acs.commons.workflow.bulk.execution.model.Workspace;
