+++ bundle/src/main/java/com/adobe/acs/commons/workflow/bulk/execution/impl/BulkWorkflowEngineImpl.java
@@ -23,6 +23,8 @@ package com.adobe.acs.commons.workflow.bulk.execution.impl;
 import com.adobe.acs.commons.util.QueryHelper;
 import com.adobe.acs.commons.workflow.bulk.execution.BulkWorkflowEngine;
 import com.adobe.acs.commons.workflow.bulk.execution.model.Config;
+import com.adobe.acs.commons.workflow.bulk.execution.model.Status;
+import com.adobe.acs.commons.workflow.bulk.execution.model.SubStatus;
 import com.adobe.acs.commons.workflow.bulk.execution.model.Workspace;
 import org.apache.commons.lang.StringUtils;
 import org.apache.felix.scr.annotations.Component;
