+++ bundle/src/main/java/com/adobe/acs/commons/workflow/bulk/execution/impl/runners/SyntheticWorkflowRunnerImpl.java
@@ -20,11 +20,9 @@
 
 package com.adobe.acs.commons.workflow.bulk.execution.impl.runners;
 
-import com.adobe.acs.commons.fam.ActionManagerFactory;
-import com.adobe.acs.commons.fam.DeferredActions;
 import com.adobe.acs.commons.fam.ThrottledTaskRunner;
 import com.adobe.acs.commons.workflow.bulk.execution.BulkWorkflowRunner;
-import com.adobe.acs.commons.workflow.bulk.execution.impl.Status;
+import com.adobe.acs.commons.workflow.bulk.execution.model.Status;
 import com.adobe.acs.commons.workflow.bulk.execution.model.Config;
 import com.adobe.acs.commons.workflow.bulk.execution.model.Payload;
 import com.adobe.acs.commons.workflow.bulk.execution.model.PayloadGroup;
