+++ bundle/src/main/java/com/adobe/acs/commons/workflow/bulk/execution/model/Status.java
@@ -18,12 +18,17 @@
  * #L%
  */
 
-package com.adobe.acs.commons.workflow.bulk.execution.impl;
+package com.adobe.acs.commons.workflow.bulk.execution.model;
 
-public enum SubStatus {
-    ERROR,
-    DEACTIVATED,
-    STOPPING,
-    FORCE_TERMINATED,
-    SLEEPING
+public enum Status {
+    NOT_STARTED,
+    FAILED,
+    STOPPED,
+    RUNNING,
+
+    ABORTED,
+    COMPLETED,
+    SUSPENDED,
+    TERMINATED,
+    FORCE_TERMINATED
 }
