+++ bundle/src/main/java/com/adobe/acs/commons/workflow/bulk/execution/model/Workspace.java
@@ -21,8 +21,6 @@
 package com.adobe.acs.commons.workflow.bulk.execution.model;
 
 import com.adobe.acs.commons.workflow.bulk.execution.BulkWorkflowRunner;
-import com.adobe.acs.commons.workflow.bulk.execution.impl.Status;
-import com.adobe.acs.commons.workflow.bulk.execution.impl.SubStatus;
 import com.day.cq.commons.jcr.JcrUtil;
 import org.apache.commons.lang.ArrayUtils;
 import org.apache.commons.lang.StringUtils;
